Round Robin Based Priority Scheduling (Time Quantum 2 seconds)
Now in our question it is a fusion of priority scheduling as well as round robin as there is a gap of 2 seconds devoted in between every context switching. Since the priority is the least number assigned so at first the program will execute that first with a time gap of 2 seconds and then it will context switch to the next process having the highest priority that process will also run for 2 seconds and in this way the operating system will perform the task and the average waiting time and average turn around time is calculated out.